feat: add increment month field and update dependencies
This commit introduces a new "INCREMENT MONTH" field to the main data entry and update forms. This allows users to specify whether an employee's increment is due in July or January.

The `month_update.dart` logic has been updated to handle automatic increment calculations based on this new field. When saving data for June or December, the system will now automatically calculate and save the incremented salary for the following month (July or January, respectively) if the increment month matches.

Additionally, the `http` and `http_parser` dependencies have been added to `pubspec.yaml` and `pubspec.lock` has been updated accordingly.

Finally, a minor adjustment was made in `itax.dart` to use the alias `uh` for `universal_html` to avoid naming conflicts.